core/busybox: create /tmp dir
the --initdb will be ignored if there are a /var/lib/apk/world file and will
not create the needed /tmp. Since the apkovl contains this file, the /tmp will
never be created, which causes busybox.post-install to fail due to the missing
/tmp. The /tmp is provided by alpine-baselayout which depends on busybox so
its not there whel busybox is installed.

As a workaround we add /tmp to busybox package. The correct way to handle this
is to make apk add create /tmp if its not there already when extracting the
install script.