main/iproute2-qos: add htb burst and fix privileged class ceiling
The HTB classifier uses a burst parameter to figure how much of
packets can be queued when the class is lending from other classes.
It should be set automatically by tc / kernel, but appears not be
set, so set it explictly. This fixes htb class not getting full
speed if difference of ceil and rate is large.

This also changes Privileged class to get ceil of the total
bandwidth instead of the previous hard limit of guaranteed rate.
While, it is consider good practices for Privileged class to have
hard ceil less than max. speed, it will cause problems links with
asymmetric speeds (e.g. ADSL 24M/1M cannot gain full download
speed if Privileged class is limited this heavily). It might be
a good idea to later have the hard ceiling configurable, but for
now it's better to ceil it to wire speed.
(cherry picked from commit 537214e7bfd9ce56e8f0c418ac0cd628365387b8)