main/vsftpd: Changed libcap-dev from an excluded make dependency to a required one.
Changed libcap-dev from an excluded make dependency to a required one,
to resolve a linkage error. The exclusion dates back to when the package
was in testing, and the version has been advanced since then. I don't
know why libcap was excluded. Before this change is adopted some more
research is required.