main/nss: re-organize the package
libnss was a bad name since it contained various libs. We put the
libs in the mai, 'nss', package and the tools in a nss-tools package.
This is how fedora does.

We also make sure that the .so symlinks are in nss package instead
of -dev package due to the way soname is.