main/vlc: unbreak streaming of files to network
The uclibc patch disabled using of monotonic clocks for
pthread_cond_timedwait(). This was needed for LinuxThreads as it
does not support it. However, this completely breaks the NPTL build.
We really need to use the monotonic clock as that's how
pthreads_condattr_init constructions our conditions. The original
patch was flawed: it should have disabled the monotonic clock via
other ways.