Initial APKBUILD for pylint
Package description:
Pylint is a python tool that checks if a module satisfies a coding
standard. Pylint is similar to PyChecker but offers more features,
like checking line-code's length, checking if variable names are
well-formed according to your coding standard, or checking if
declared interfaces are truly implemented, and much more.
The big advantage with Pylint is that it is highly configurable,
customizable, and you can easily write a small plugin to add a
personal feature.