Initial APKBUILD for nebula
Package description:
Nebula is a network intrusion signature generator. It can help
securing a network by automatically deriving and installing
filter rules from attack traces. In a common setup, nebula
runs as a daemon and receives attacks from honeypots. Signatures
are currently published in Snort format.
The code was written to be fast. A signature is not of much value
if the generation process takes hours or days. With nebula, you
should get a first revision within a few seconds. As more attacks
of a kind are submitted, signatures get better and nebula publishes
updated revisions.