Initial APKBUILD for etherdump
Package description:
etherdump is an extremely small packet sniffer. It can be used to
dump raw ethernet frames for parsing (optimally with Ethereal's
text2pcap), or for viewing the network traffic of a few common
TCP/IP protocols. On a uClibc system a dynamically-linked version
is about 8kB and a statically linked version is about 30kB because
it does not depend on libpcap.