Initial APKBUILD for  fabric
Package description:
Fabric is a Python (2.5 or higher) library and command-line
tool for streamlining the use of SSH for application deployment
or systems administration tasks.

It provides a basic suite of operations for executing local or
remote shell commands (normally or via sudo) and uploading/
downloading files, as well as auxiliary functionality such as
prompting the running user for input, or aborting execution.

http://www.fabfile.org