Initial APKBUILD for nwipe
Package description:
The nwipe is a command that will securely erase disks using a variety of
recognized methods. It is a fork of the dwipe command used by Darik's
Boot and Nuke (dban). Nwipe was created out of need to run the DBAN dwipe
command outside of DBAN. This allows it to use any host distribution which
gives better hardware support. It is essentially the same as dwipe, with
a few changes:
- pthreads is used instead of fork
- The parted library is used to detect drives
- The code is designed to be compiled with gcc

Website: http://nwipe.sourceforge.net