Initial APKBUILD for ms-sys
Package description:
This is a Linux program for writing Microsoft compatible boot
records. The program does the same as Microsoft "fdisk /mbr"
to a hard disk or "sys d:" to a floppy or FAT partition except
that it does not copy any system files, only the boot record
is written.

Website: http://ms-sys.sourceforge.net/