testing/py-pip: new aport
pip is a tool for installing and managing Python packages, such
as those found in the Python Package Index.

pip is a replacement for easy_install. It mostly uses the same
techniques for finding packages, so packages that are easy_installable
should be pip-installable as well. This means that you can use pip
install SomePackage instead of easy_install SomePackage.

http://www.pip-installer.org