testing/tcpreplay: new aport
Tcpreplay is a suite of  tools for UNIX  operating systems which gives you
the ability to use previously captured traffic in  libpcap format to test a
variety of network devices. It allows you to classify traffic as client or
server, rewrite Layer 2, 3 and 4 headers and finally replay the traffic
back onto the network and through other devices such as switches, routers,
firewalls, NIDS and IPS's. Tcpreplay supports both single and dual NIC
modes for testing both sniffing and inline devices.

http://tcpreplay.synfin.net/