testing/tcpflow: new aport
tcpflow is a program that captures data transmitted as part of TCP
connections (flows), and stores the data in a way that is convenient
for protocol analysis and debugging. Each TCP flow is stored in its
own file. Thus, the typical TCP flow will be stored in two files,
one for each direction. tcpflow can also process stored ‘tcpdump
packet flows.

https://github.com/simsong/tcpflow