main/abuild: more provides fixes
need special handling when file name is .so (no version number suffix)
and soname has a version number suffix. This is so we dont set provides
version to 0 when it should have been something else.

This fixes issue with libgcj/gcc-java who ships a file libgcj_bc.so with
soname libgcj_bc.so.1 in addition to the libgcj_bc.so.1.0.0 file.

We cannot have 2 provides with same name but with different version.