main/nfs-utils: fix for wrong getaddrinfo assumptions
The nfs-utils assumes getaddrinfo(AI_NUMERICHOST) never fills in
ai_canonname, but it appears to do on uclibc. It also assumes that if it
filled in, it will be allocated separately and that freeaddrinfo() will
free it. This is not what uclibc does.

This patch should fix a potensial memleak on nfs server and probably it
will fix nfs server on x86_64

ref #1486