Fixed postfix bug causing postconf -m to segfault
Running postconf -m resulted in a segfault and no output.
The bug originated in dynamic-maps3.patch. postconf_other.c was modified by
this patch such that it called concatenate(), but the file was not modified to
include <stringops.h>, resulting in the function being undeclared. This
resulted in gcc incorrectly assuming a default prototype for it with return
type int instead of char*, causing pointer truncation on 64-bit systems,
causing a segfault when the truncated pointer is passed to another function.
The fix is to #include <stringops.h> in postconf_other.c. The patch has been
updated.

I don't have abuild 2.14 on this machine so I'll leave updating the SHA256
checksums to whoever commits it.