main/syslinux: add reboot entry
We don't install /boot/reboot.c32, but if it's present (it can be copied from
/usr/share/syslinux/reboot.c32), we add a menu entry for it---just like
with memtest.

Add a comment to /etc/update-extlinux.conf stating this.

It'd already be possible to get a reboot entry using the
/etc/update-extlinux.d/ folder, but this patch provides a more intelligent
framework with nicer layout.

The syslinux sourceball also provides a shutdown module, but I couldn't get this to
work and looking at the sources reveals it to be for machines with APM enabled.
Not sure how many machines that applies to anymore.