main/gcc: disable ifunc attrs, dont disable symvers for uclibc
Seems gcc 4.8.1 started using ifunc attributes in libatomic which go
uses heavily. This requires ldso support which is not done in uclibc
nor musl. Only in glibc at the time. GCC configure does not detect it
properly.

Also, do not disable symvers for uClibc - otherwise we break ABI for
libstdc++