abump: error out on first faild package
If some package fails, then stop processing the rest. This is so we
don't risk test build the rest for the packages against wrong lib.

For example, if 'abump libfoo-2.0 foo-2.0' fails on libfoo, we don't
want testbuild foo-2.0, which might end with success against libfoo-1.

We also echo the list of packages that was not bumped, so its easy to
copy/paste next run.