main/gcc: fix ada bootstrap once more
The new gentoo piepatches add -fstack-check to defaults which make
ADA bootstrap break on certain platforms (at least arm).

The configure patch affects stage1_cflags, which is needed if the
currently installed host compiler has -fstack-check in it.

The Makefile.in part affects rest of the bootstrap stages and is
required if the current patch set includes -fstack-check in the
default spec.