Patch for PSAD to fix sending mail & service starting.
Added start() to /etc/init.d/psad - it was not starting without it.

Added a dependency for mailx to APKBUILD & fixed /etc/psad/psad.conf setting for mail (ssmtp on it's own was not enough - mail was
incorrectly formed).

Tested on a live server - scans are detected, blocked & email notifications sent.