main/fakeroot: hide unharmful dlsym errors
this got uncovered when we enabled libacl support. the acl_* symbols
are tried to be resolved but they will not exist in the app, if it's
not linked against -lacl. it's perfectly ok for these to not be found.

(cherry picked from commit e56db141ce09c403f59621b4e56a8025a387ef6c)