main/squid: allow more than 16384 filedescriptors
It seems that when squid is compiled with --with-filedescriptors, that
becomes a hard limit[1]. Without this compile option it will use the
system ulimit setting, which is the behaviour we want.

For compatibility we set SQUID_MAXFD to 16384 if unset but still allow
it to go over 16384 if needed. It should also be possible to set
max_filedescriptors in squid.conf[2].

fixes #2850

[1] https://bugzilla.redhat.com/show_bug.cgi?id=976815
[2] http://www.squid-cache.org/Versions/v3/3.3/cfgman/max_filedescriptors.html