main/tinc: fixed init scripts
This patch fixes an init script bug which caused 'restart' to fail
& which sometimes could only be cleared by a reboot (--localstatedir
was set as /usr/var so the pid was never found).

I changed 'stop' to use the start-stop-daemon as in tinc 1.1pre10 there
is no longer an option to --kill. Message for an already stopped daemon
corrected.

For the same reason I also changed 'reload' to use kill -1 $pid.