main/musl: remove post-install and use only post-upgrade
The problem is that /bin/sh depends on bash which depends on musl so we
get a circular dependency. To solve it we simply remove the
post-install. The problem it is supposed to fix is very unlikely to
happen on new installs and if it does, it will be resolved with next
upgrade.