main/php: fix php-fpm script
ref #2685

The problem was that it usesd start-stop-daemon --exec but the argv[0]
name didnt correpond so start-stop-daemon could not know that it was up
and running.

The fix is to use pidfile.

Script is based on update from gentoo.

(cherry picked from commit 0449861a8f541e2091c0e4b049682f8deee67333)