testing/sword: add new package
Sword is a library for free Bible software.
It can also be used for other books.

sword-utils contains a number of tools, including installmgr to add "modules"
(books), diatheke to display them, and numerous file format converters.

sword-lang contains config files to translate the names of the books
of the Bible from English into various languages.