main/slony1: fix logrotation
we fix logrotation by only supporting syslog and not redirect the slon
output to a given logfile.

This is becase when redirecting, the only way to do log rotation is restart
the service to reopen the log file. We do not want force restart of service
which will close open connections.