main/busybox: fix coldplug with modprobe -a
We call 'modprobe -a <alias> ...' from initramfs init for coldplugging.

The upstream commit 4e314faa0a (modprobe,rmmod: reject module names with
slashes) introduce a check for slashes in the modprobe name. This broke
loading of module aliases with slashes which are often found in dmi
modalias.

We fix this by allowing slashes when -a option is given.