main/fping: enable suid root for fping6
It should be possible to make ping work as normal user with recent
linux: https://lkml.org/lkml/2011/5/13/382

However, it appears that it requires relatively intrusive changes in the
ping utility. Its not just opening the socket as SOCK_DGRAM and think
everything is ok (like it appears to work on OSX)

Busybox tried to add support for it:
http://git.busybox.net/busybox/commit/networking/ping.c?id=72745632a13ccd12232127b31e1656f2f7ebcaff

But later reverted it:
http://git.busybox.net/busybox/commit/networking/ping.c?id=f0058b1b1fe9f7e69b415616096fb9347f599426

The patch for the normal ping utility is also relatively intrusive:
http://openwall.info/wiki/_media/segoon/iputils-ss020927-pingsock.diff

I could also not find anything that indicates that this works for ipv6
in kernel. There are no /proc/sys/net/ipv6/ping_group_range

So it looks like we need to set suid root for now.

ref #3000

(cherry picked from commit d37ee34b7691ca0e2d59f39ec2dbc9484105c087)