main/llvm: add support for polly
additionally, to make it easier to take advantage of polly's functionality, it
has been statically compiled into the tools themselves rather than as a plugin
that needs to be manually loaded before use.

polly's GPGPU codegen functionality has sadly been disabled in this build.
it appears that things are supposed to work without CUDA available, but this
doesn't seem to be true.