testing/davfs2: new aport
http://savannah.nongnu.org/projects/davfs2
Linux FUSE (or coda) driver that allows you to mount a WebDAV resource

I based this build off of the Gentoo ebuild and the OpenWRT build files/patches
https://github.com/openwrt/packages/tree/master/net/davfs2

Had to modify one of the patches to add iconv.h to automake tests due to HAVE_ICONV_H
being used but HAVE_ICONV was defined with 'iconv' rules and both definitions being used.

Regarding maintainership - I could possibly keep up with the project and keep it up-to-date.