main/busybox: enable support for syslog.conf
The feature is pretty useful for users (like me) who would like to
organize their logs by using facilities and since it doesn't increase
the size of the syslog binary significantly I thought that it is worth
it to enable it by default.

For users who don't want to use this feature nothing changes since the
feature is only used when /etc/syslog.conf exists. Busybox doesn't
complain when that file doesn't exist so I decided not to create it by
default.