main/mdocmal: don't rebuild the mandoc.db from scratch every time
mdocmls makewhatis(8) implementation should be able to build the
mandoc.db incrementally. Removing it forces mdocml to regenerate it
completely which can take a lot of time if you have a bunch of man pages
installed on your system and thus slows down package installation.

Furthermore, I also switch the string encoding from ascii to utf-8. In
addition to that I noticed that some files weren't listed in the source
APKBUILD variable.

Seem to work for me but further tests would be appreciated.