git-manpages: remove package, replaced by git-doc
First of all git-manpages is outdated and furthermore all man pages
shipped by this page are also shipped by git-doc except for the fact
that the git-manpages man pages are not compressed.

If you disagree with the removal of this package than at least make this
package a subpackage of the git package.