main/mdocml: use a cron instead of a trigger to invoke makewhatis(8)
The database created by makewhatis(8) doesn't need to be up to date all
the time, it is not critical at all when the database doesn't contain a
newly added manpage for a few hours. So instead of slowing down package
installation by using a trigger the makewhatis(8) database is now
created and updated using a daily cron. The daily cron is run once after
the first install of mdocml as well.

Many distribution (e.g. Gentoo) also use a cron instead of a trigger to
invoke makewhatis(8), in my opinion it is simply saner to it this way.