main/hub: upgrade to 2.2.1
hub has switched from ruby to go for performance reasons which leaves us
with the usual go dependency tracking problem. I decided to simply
invoke their build script which has the huge disadvantage that the
integrity of the dependencies cannot be verified.

Also remove Eivind Uggedal as a maintainer since this package was
outdated in a major way and he didn't touch it for an entire year. I
don't want to maintain this package because I don't use it anymore. I
would suggest that somebody else takes over the maintainership for this
package and finds a solution for the problem above or alternatively I
would simply move this package to unmaintained/ or testing/ until a
solution for the problem above and/or a new maintainer is found.