main/go-boostrap: new aport
This is a more minimal version of the main/go aport. It is only supposed
to be used for bootstrapping go1.5 and to clarify this point it is called
go-bootstrap instead of go1.4.

This package doesn't ship a tools or cross subpackage anymore because
doing so would increase the amount of redundant 'code' in this APKBUILD
and neither tools nor cross compilers are needed for bootstrapping
go1.5.